fixes test errors due to #263
